+++ src/defaults.ts
@@ -1,4 +1,4 @@
-const WS_TIMEOUT = 30000;
+const WS_TIMEOUT = 10000;
 
 export {
   WS_TIMEOUT,